Refactor Types and Constants
